Remove pointless (slow) code
This code was originally designed to show the email address for a 'specific' group.
E.g. if you contacts have an email address tagged under 'Quicksilver', and your 'preferred group' was set to
'Quicksilver' (which was the default), then that email would have been shown.

In reality, it just accounted for 40% of the slowness of address book record indexing